+++ src/matrixfree.jl
@@ -22,8 +22,8 @@ function _apply_Δₓ₄!(y::AbstractVector, x::AbstractVector, h₁) # 1D Lapla
     end
 end
 
-# L = -Δ  + R    
-function _apply_Lₓ₄!(y::AbstractVector, x::AbstractVector, h₁) # 1D Laplace operator (4th order) with periodic bc
+# L = -Δ + R
+function _apply_Lₓ₄!(y::AbstractVector, x::AbstractVector, h₁) # 4th order, periodic bc
     nx = length(x)
     length(x) == length(y) || throw(DimensionMismatch())
     Σx = sum(x) / nx
@@ -38,7 +38,7 @@ function _apply_Lₓ₄!(y::AbstractVector, x::AbstractVector, h₁) # 1D Laplac
 end
 
 ### Constant Nullspace Projection
-# if 1 ∈ Δ, then Δϕ = ρ is not well posed but (Δ + R)ϕ = (1 - R)ρ is. 
+# if 1 ∈ ker Δ, then Δϕ = ρ is not well posed but (Δ + R)ϕ = (1 - R)ρ is.
 function _apply_Rₓ!(y::AbstractVector, x::AbstractVector) # Nullspace projection
     nx = length(x)
     length(x) == length(y) || throw(DimensionMismatch())